WT-13187: Ensure that initial scripts are properly closed
When serving the main page, either set up through the hybrid or
conventional boot, the script will run, ensuring a response it sent to
the server. This will contain all correct initial start-up information,
and the way the follow-up requests ought to be handled.

In this logic, we also append all initially defined resources for the
application, like stylesheets or JavaScript. The latter is loaded via
WApplication::require. Any script added here was not properly closed.
While the script was still properly loaded, the actual bootstrapping
mechanism that was below it would be commented out, and not be
interpreted by the browser. This results in every new request coming in
to actually be considered a NEW request, and not part of the same
application.