WT-13075: Add option to popup auto positioning
Currently, the behaviour of the automatic positioning used by popups
make them position over their anchor in case there is not enought place
above and bellow that anchor. This may however not always be desireable,
like in the case of WSuggestionPopup which will hide what the user is
writting.

This add options that allow to choose whether the popup should
prioritize showing itslef completly or prioritize not hidding the anchor
by going over it.