WT-13481: Track cookies added by the `WApplication`
When the initial request to set up a user session on the serverside is
sent to the server, some information regarding the client is sent along.
This state is accessible throughout the application as the
`WEnvironment`. This is seen as a "read-only" informative bundle.
However, it is likely that some of its information changes during the
application's lifetime. One such item are cookies. When logging in,
remembering information, ... cookies are often used.

To ensure that these cookies can be correctly caught by the application,
they are remembered by the application. This concerns cookies that are
added within the same `WApplication`. If the user refreshes, or starts a
new `WApplication` in any way, the list is cleared, but will now be
correctly appended to the `WEnvironment`.