WT-13481: Make sure logging out a user also removes serverside cookies
This fixes an issue where a removed cookie will be correctly removed
from the client's browser, but not from persisted storage on the server.

This used to have the effect that older cookies that were not renewed,
but removed from the database can still be used to gain access to the
user's session. The cookie and its value would have to be manually added
to the browser, and the server would detect it, and be able to use it.