WT-13501: Do not make redirected processes use inline handlers
Processes like SAML or OAuth that will result in a redirect did so with
a very simply JavaScript bootstrapper, that performed the redirect upon
loading a basic HTML where an inline function would be added to the
body's loaded handler. A default CSP (which omits `unsafe-inline`) does
not allow this. This has been replaced with a pure JavaScript
equivalent.