WT-13501: Make all `DomElement::setEvent` render as JavaScript only
As with the previous commits, we do not want inline JavaScript handlers.
Instead we prefer pure JavaScript to define the functions, as to not
violate the default CSP.

Any event added to the `DomElement` with `setEvent` now correctly
renders out as JavaScript, with a fallback in case the `id` is not yet
known.