WT-13507: Only propagate the parent enabled-state if the widget is not disabled
Moving a disabled widget to an enabled parent would sometimes, wrongly, make it
enabled again. In case a disabled container, with passively disabled
widgets (inheriting the state from their parent), would be moved to and
enabled container, the widgets would be made enabled again. The
propagation of the enabled state did not correctly check the state of
the parent widget.