WT-13507: Add tests for disabled widgets and interactions
The tests dictate the expected behaviour, namely that disabled widgets
ought to be both marked as disabled, and should be visually styled as
such (with the "Wt-disabled" class, or similar, depending on theme).

Children of disabled widgets should NOT be marked disabled themselves,
but are visually marked as such, with the above styleclass.

Any change to the parent is propagated through to the child.