WT-13517: Reset `WContainerWidget` layout flags
When either the layout changes or is removed, flags are set to ensure
that in the next rendering step the `WContainerWidget` correctly know
what to do. In some cases it may need to remove all its children, and
then readd them.

Due to a change in WT-13261 (c0ad9dc1103b3555c2cf92cf6f1a4be46dcbc54a)
a flag was set to let the renderer know the layout was reset. However,
if this container was rendered not fully, but as part of a bigger
update, the rendered state is not propagated, and thus the flag is not
cleared. In this case, the flags need to be manually removed, at the end
of `createDomElement()`.