WT-13521: Delay encoding of WTemplate rendered text
Due to a change in WT-13501, the RefEncoder will spawn dummy objects,
and use their generated IDs to add click handlers to links. In case this
was rendered inside a WTemplate (using setInternalPathEnabled(true)),
but hidden inside a condition, the JavaScript would be rendered, but of
course no object could be found.