WT-13550: Fix wrong href to local path
Fixes a bug where the RefEncoder would change the href, making it point
toward the current path. This was not impacting the navigation in most
cases since it did not impact the click handler.

This was solved by removing a fix for a bug that we cannot reproduce
anymore. This bug would lead to pages being reloaded due to having the
click handler being called to late when clicking on a link.