WT-13567: Correctly schedule recursive reading for `SessionProcess`
The `read()` of a `SessionProcess` is called recursively, in that after
each read event, we reschedule the operation. Here however, it is
wrongly assumed that both the socket and context are still active.

This edge case would surface when serving very small responses (e.g. a
small static file). As the child process would essentially immediately
be terminated, and the socket closed. This lead to the `read()` being
called on a closed socket.