WT-13499: "Remove" unusable Cookie ctor from jwt
Removes the Cookie constructor using WDateTime and the one using
std::chrono::seconds from java target. This was not usable in java since
the Cookie class is not translated to java, but this was not visible
when reading the c++ code. This makes it clear that those constructor
should not be used for code that will be translated to java.