WT-13600: Fix CSP violation when avoiding redirect
Removes inline javascript in href of WAnchor and WAbstractArea with no
link. This was preventing tab chain from skipping the element while
disabling redirection when clicking on the WAnchor or the WAbstractArea.

To keep the same behaviour, we add the new Wt-no-default class attribute
to the DOM element which disable the default action of the browser for
all of the event of the widget, and we set the href attribute to #.