WT-13553: Explicitly convert to bool for `email_verified` in OIDC
In case a string was received here that contained `true` or `false`,
the JSON framework would detect the value being present, and try to
retrieve it as a `bool`. However, it should first be made a boolean
explicitly by calling `toBool()`, before the `orIfNull(false)` triggers,
which would lead to a `TypeException`.

Bug patched by Stephan Kaiser