WT-13621: Ensure that QrCode.hpp is installed
TotpQrCode requires this file. It has to be installed in the include
directory if we want to allow developers to use the class. Additionally,
we need to add WT_API to export the symbol for shared libraries.

It's reasonable to allow this, since developers might want to allow
users to setup MFA with a QR code from within their application, and Wt
currently doesn't offer an alternative.