WT-13626: Process MFA environment before render
AuthWidget::processEnvironment should be called initially to process
environmental information relevant for authentication. The login state
can become RequiresMfa if the user has a valid auth token and also has a
Mfa step. Checking whether the user also has a valid auth token is
delayed until the widget is actually rendered, as part of createMfaView.
It seems more appropriate that processEnvironment also checks for a
valid MFA token. This ensures that the widget can immediately render a
logged in view if there is such a valid token, thus skipping the
creation of the MFA view.