WT-13585: Emit the `currentWidgetChanged()` signal only once
This is only performed when the widget in the `WStackedWidget` actually
changes. We also keep track of whether the emit has happened. This
should be done for Lazy widgets. These are not always yet present when
the `WMenu::select()` happened. But the `select()` call where their
contents have been loaded will happen a little later.