WT-13478: Change checkboxes in Auth templates
The checkbox layout when using bootstrap were not used like they ought
to be. Normally, the label of a checkbox is behind it, not in front.
This is especially important in Bootstrap 5 as having the label in front
of the checkbox causes it to not be colored properly. For that reason,
the label of the checkboxes in the auth forms are now displayed behind
the checkboxes in all versions of Bootstrap.

To keep the help-text separated from the label, the help-text of the
checkboxes are now displayed under the label, and so is the case of the
help-text of the other widgets in the auth form for consistency.