WT-13536: Fix WTableView loading bug due to lag
When using WTableView, when the client scrolls outside of the data
buffered in the browser, the "scrolled" signal is emited, which will
request the correct data to the server. Once those new data are
recieved, the browser updates the cache, removing the data that are not
visible anymore with the one that are visible.

However, scrolling back to the buffered data before the new data arrived
would not emit a new "scrolled" signal. Due to that, once the new data
arrived, the old data (which is what the user sees) would be deleted,
and would not be reload until a scrollevent or a window resize happened.

This makes sure that the "scrolled" signal is correctly emited in that
situation.