WT-13536: Fix QueryModel cache not always full
In case the row to be retrieved from a QueryModel does not have
enough rows following it, the cache would not be completely filled. Now,
the cache will always be filled with exactly the number of entries in the
batch size. If not enough items follow the queried row, a number of rows
preceding the queried rows are appended.