WT-13685: Do not loop resizing too much
In case of stretchable content in a StdGridLayout, we calculate the
available space, and size the widget according to it.

In case of a small space left (less than a pixel). The logic will keep
iterating, even though no resize can take place. The loop should ensure
that the available space is large enough for resizing to occur.