WT-13670: Fix error with spacing in flex layout
Fixes the use of horizontal spacing that was used in place of vertical
spacing for FlexLayoutImpl. This was not causing any issue however since
only WGridLayout makes the distinction between horizontal and vertical
spacing, which is also the only layout never using the flex layout
implementation.

The horizontal and horizontal spacing should now be used correctly.