WT-13670: Fix ignored border width
Fixes borders width being set to 0 in case their width was given using
"thin", "medium" or "thick". This also has for side effect to solve
an issue with layouts using JS inplementation. The issue caused spaces
to be added between the layout items and the borders, and would mainly
happend when resizing or zomming.