WT-13699: Add the `Wt::TextDecoration::None` option
This will function different from `Wt::None` when setting the text
decoration of a widget.

The `Wt::None` flag will empty out all text decoration, and essentially
render out as empty. This makes its behaviour the "default". Meaning
that it can still be affected by other CSS rules, if they set the text
decoration.

On the other hand, when setting `Wt::TextDecoration::None`, will it
actually render out the `none`. And that will apply to the element,
stopping other CSS rules.

For example a global rule on `span` will apply to all `WText` items, but
setting the `Wt::TextDecoration::None` of a specific item blocks the
former rule.