WT-13701: Only select the `WStackedWidget` item if it is valid
When a `WMenu` selection takes place, and a `WStackedWidget` is used to
display the items, first an unselect takes place, followed by the
selection of the correct widget.

This unselect call, always uses the index `-1`. The `WStackedWidget`
did not check for this specific index.

This was present in 455a9534, but undone in 6140bc9e, due to
simplifications in `WMenu`. However, this case was not added to
`WStackedWidget` again.