WT-13670: Fix items in flex layout not resizing
When using JS layouts, the width and height of an element are used as
preferred width and size, but does not constrain the final size of the
element, which will be resized in order to take all of the space
attributed to it.

Flexboxes only behave this way for the width when its flow is "row" or
"row-reverse". In case the flexbox flow is "column" or "column-reverse",
te width attribute of any of it's child will be the actual width of that
child. For the height, it is the same, but when the flow is "column" or
"column-reverse" instead of "row" and "row-reverse".

In order to solve this issue, the items inside of a flex layout are now
wrapped in another flexbox with a perpendicular flow (row <---> column).
The wrapper will have its "flex-basis" attribute set to the width or the
height of the element (depending on the flow of the layout), and the
element will have that same value set to auto, making it fill the whole
wrapper, solving the issue.