WT-13670: Justified item in flex layout shrinking
When having multiple items justified in a flex layout, if the sum of
their size were bigger than the size of the container, the items could
overlap. To solve this issue, the items justified in flex layout will
now have a shrink factor of 1, allowing them to shink if necessary.