WT-12424: Consistant rendering of EventHandler
The rendering of EventHandler was different depending on whether nonce
were used or not. This was unnecessary since now the EventHandlers are
never set using inline JS anymore. Both cases are now handled in the
same way.

This seems to fix the crashes happening when a WAbstractMedia was is a
WTemplate, when there was also a large WTableView. The crash was due to
the WAbstractMedia JS function not being added to the DOM element. Which
for some reasons was only happening when not using nonce.