WT-13712: Add styling for anchors in `WTableView`
In case a `ItemDataRole::Link` is provided in the model, the item is
rendered as an anchor element. This element did not get the same styling
as the `div`. Since for Wt, an anchor is a `WContainerWidget`, it was
simpler to add the styling rules, then to wrap it in an additional
container.