WT-13569: Allow to force apply theme style
Most of the time, when WTheme::apply() is called on a widget, it should
only be applied when the theme style is enable for the parent widget.
This is however not always the case, since sometimes this makes changes
that are mandatory for a feature of the parent widget to work. Most of
the time, these call are made during the construction of the parent
widget, which effectively force the style to be applied since by
default, the theme style is enabled on all widget.

However, this forces all of those widget to be created at the
construction of their parent, which is not always desirable. For this
reason, this adds the possibility to force  the theme style to be
applied to a widget.