WT-13569: Fix dangling pointer in WToolBar
In case WToolBar has multiple groups, it keeps a pointer to the last
group in order to be able to add new widgets to it. When removing a
widget from the WToolBar, if that widget is the last widget of its
group, the group is removed. In case the group removed was the last
group, the pointer kept by the WToolBar still pointed to that deleted
group.