WT-13569: Allow to delay appling theme style.
Setting the theme style should only be done during rendering unless it
is forced since before rendering, whether the theme style should be used
or not is still unknown. Delaying applying the theme style can however
sometimes be quite inconvienient as it requires to keep track of every
widgets needing special theme styling, as well as the role it has,
making the code more complex.

This ought to facilitates delaying the application of the theme style
by adding the possibility to schedule it for the end of the rendering
of the widget (or earlier if needed).