WT-13733: Specify an QueryModel::createIndex override
This is necessary because the `WStandardChartProxyModel` will assume an
index of `-1, -1` as its fallback.

For all other models this is not an issue, since the
`WModelIndex::isValid()` check will correctly catch this, since it will
be considered invalid. For a `QueryModel` this is not the case. Checking
the index, means implicitly creating it, resulting in a valid index.

The `WAbstractItemModel::createIndex()` is now `virtual` so the
`QueryModel` can override it, and perform the additional bounds-check.

Note: this check is only performed in the negative direction. Meaning
indices below 0 are considered invalid. The higher bound (`rowCount()`
or `columnCount()`) is not checked. This to ensure that inserts still
can correctly occur, and to not invoke the count query too often.