WT-13569: Fix style class duplication.
DomElement::addPropertyWord() checks that the word added are not already
present. This does however not work when multiple words are added at the
same time, potentially leading to having words being duplicated. To
avoid this duplication, DomElement::addPropertyWord() should not be used
to add more than 1 word at the time, unless we are sure that every words
added are not already present in the property.