WT-13729: Rename Json::Object::isNull()
Json::Object::isNull() (with no parameters) is a function used only for
jwt, in which it is translated to "$object; != null" which effectively
checks that the object is not null. Which makes the name very confusing.

Since this function is used to check that the Json was parsed without
any error, it is now renamed Json::Object::isValid() for clarity.