WT-13777: Do not allow model deletion propagation upon invalid indices
It was possible for an invalid index to prompt the model to change
itself. This should not be the case, as the change itself was invalid.

Since `WTreeView` can have arbitrary models, the upper bound is not
validated, only the lower bound. Otherwise, if using a database model,
this can have a count statement as a side-effect.