WT-7986: Stop sending unchanged form data
The form data of each form object were send every time a signal was sent
from the client to the server. This means that it was common for the
same data to be sent over and over again, especially when using big
WTableView.

While this did increase the size of the messages sent by the user to the
server, it also allowed the server to know the exact state of the client
when the message was send, which may not include the last updates sent
by the server to the client. Therefor, simply avoiding to resend the
form data that did not change is not sufficient.

This resolves both of the issues by having the server keep in memory each
of the last form data sent by the client, allowing the client to not
resend the data that were not changed since the last signal sent. The
client also resends all the form data in a signal sent after receiving an
update from the server, to update the cache of the client with the
new values set by the server.