WT-7986: Add max form-object resend
To mitigate potential increase in server update message size, due to
asking to resend to many form-object, a limit is added to the number of
id of form-object to resend sent to the client. If this limit is
reached, the server will instead ask the client to resend all form-data.