WT-13818: Fix WGroupBox removing flex layout
In order to resolve issue 11882, the flex-direction of the WGroupBox was
set to column. This did however override any flex layout set to the
WGroupBox when changing its title.

To solve this issue, WGroupBox now has a WVBoxLayout that is used to
solve issue 11882 instead of directly changing the flex-direction of the
WGroupBox.