WT-13880: Stores EntryPoint in shared pointers
It was not possible to remove entry points while the server was running
if Wt was multithreaded. This seemed to be mainly due to the fact that
EntryPoints where stored in a queue and were past around using pointers.
This meant that removing an EntryPoint would make all of the pointers to
itself dangling pointers.

This solves this issue by storing and passing the EntryPoints in shared
pointers.