WT-13880: Improve entry point removal
When an entry point was removed, the PathSegment tree was recreated from
scratch. This is not really efficient, but it was not an issue as
removing internal path was mainly happening when the server was offline.

Now, the PathSegment that are not used anymore are removed from the tree
when an entry point is removed, which should be faster than recreating
the tree from scratch.