WT-13880: Stop serving generated resources to bots
Bots do not have any session linked to them as it is destroyed after the
page has been served to save memory. This does however lead to some
issues when bots tries to load an resource generated by the session.
Since the bot has no session, it creates a session for that same bot
again, that is destroyed after serving an empty html page.

To solve this issue, when generating the link for private resources, it
is generated as if it had the "bot" session id, for which the
WebControler will not create a session and send an error message.