WT-13880: Allow to set bot resource id
Every time a new bot loaded a page, each images on that page would be
stored as as a new public resource. To avoid adding so many identical
resources, we add allow to define the id of the resource, so that the
same resource is used for all the bots.