WT-13878: Do not allow bot agents to send session-related requests
A bot will have their session immediately terminated. Due to various
reasons (like previous indexing), they may still have session-related
paths in their indexed database. These are requests on a certain
endpoint, which have a query parameter that contains `wtd` or `signal`.
The `wtd` is the session ID, and the `signal` refers to an exposed
signal in the `WApplication`.

Of course the `WApplication` has been terminated already, so bots should
not be allowed to access them. Previously this resulted in the behavior
that these parameters were essentially ignored, and a completely new
page was served (the page on the specified path, without the
parameters).

We have noted that for certain bots this is not desirable, as they may
get "stuck" in a loop, and keep querying the same (set of) endpoint(s).

To avoid this, any request from a bot that contains session-related
information is now blocked, and a 403 is served.