WT-13406: Change logging in WWebSocketResource
Changes the log logged when a new connection to a WWebSocketResource was
opened. It now also gives both the id of the connection and of the
resource. Its logging type is also changed to "debug" instead of "info".

Additionnaly, a debug log is now also logged when the connection to a
WWebSocketResource is closed.