WT-13478: Make bootstrap_theme.xml more generic
The templates in bootstrap_theme.xml were designed in such a way that,
most of the time, the same block of text would work for all versions of
Bootstrap. This meant that often, an element would have classes that
would not work for the version of Bootstrap currently being used just so
that same block of text would also work with the other Bootstrap
versions. Furthermmore, this was in itself, hard to achieve.

This make uses of templates functions to make the templates change
depending on the version of Bootstrap used.