WT-13825: Make polyline a map item
In WLeafletMap, it was possible to add a polyline using the addPolyline
function. This polyline was however not a map item, like Popups,
Tooltips and Markers, and it was impossible to remove or update it once
it was added.

This make polyline a map item, allowing it to be modified and removed.