WT-13825: Make circle a map item
In WLeafletMap, it was possible to add a circle using the addCircle
function. This circle was however not a map item, like Popups, Tooltips
and Markers, and it was impossible to remove or update it once it was
added.

This make circle a map item, allowing it to be modified and removed.