WT-12815: Add precisions to Form::Dbo doc
In Form::Dbo, some assumption made by the code were not always clear,
and some were not even mentioned. This made it hard to figure out how to
correctly use those classes without reading how it was implemented.

This clarifies those assumptions while also adding some references in
the documentation in order to make it simpler to understand.