WT-13961: Correct `WTextEdit` clipboard pasting
When pasting from the clipboard, the JS code will verify that an image
is copied, and pasted, and include that as a data URL.

This functionality wrongly checked the `event.clipboard.types`'s to
verify the `type` field of the `DataTransferItemList`. However, this is
accessed via the `event.clipboard.items`.

A previous refactor caused this to be wrongly implemented.