WT-13970: Kill suspicious GET requests with session IDs
A GET request coming from any source, should never contain a `wtd`
parameter, unless the session was created before. If this comes in, and
is detected as a new session, the case will be handled much like a
regular bot request (see WEnvironment::agentIsSpiderBot()).

In the wild we've seen this happen, where (malicious) bots have scraped
pages, with `wtd` parameters. They then request these, with a non-bot
agent. The sessions created for them would remain alive for the regular
timeout period, taking up memory.

We now ensure that we handle these cases more aggressively. The session
is handled like a bot session. This renders out the page without any
session information, and afterwards terminates the session.

Should a real user come to this scenario, they will be granted a real
session, when they interact with any element on the page that results in
navigation.