WT-13970: Ensure a JS redirect filters any session ID out
In the rare case that a session ID be part of the URL (in the form of a
`wtd` parameter), they should be filtered out.

Specifically this happened when loading a page, that uses progressive
bootstrap, and immediately killing the session. This is done in the
previous commit, where suspected bots (identified by a weird GET
request, with a `wtd` parameter, when JS is enabled) would have their
session immediately terminated.

Upon this termination, the `request=script` request, present in the
`Boot.js` file would still be pending, and be received after the session
has been killed. This would lead to a request detected from a dead
session, causing a reload to be sent over. This reload would contain the
wrongly present `wtd` and thus remain looping over these two steps.

The `wtd` parameter is filtered out, ensuring this loop does not take
place.