WT-13367: Split URI in URL and data URI
For some paint device, data URI are handled differently than URL which
may even not be accepted. However, having DataInfo only provide a URI
forced to choose between a data URI and a URL, which could in some case
be limiting.

This replaces the uri() function from WAbstractDataInfo by the two url()
and dataUri() functions to solve this issue.