JWT-13987: Fix issue with empty relative url
It was possible for WLink::relativeUrl() to return an empty string if
the internal path was empty and both the applicationName_ and
publicDeploymentPath_ were empty. In that case, having an WLink of
LinkType::InternalPath with an empty internal path (therefor a link to
the base application path) while being on the base application path,
would return an empty string.

That would lead to crashes when a WAnchor with that link would be
rendered.

Since when it happens, the WLink is essentially referring to the page
that is currently being rendered, WLink::relativeUrl() will now return
"#" instead of an empty string in that specific case.