WT-13878: Correctly use `DEBUG_JS` for all minified scripts
There are a couple JavaScript scripts included in the CMake process,
since they are added to the "source". This mainly involves initially
sent over scripts that will load the framework, or embed the correct
HTML for Wt to set itself up.

Here, the wrong check was performed, leading to a minified script always
being served if the build was a debug build. Now, it correctly considers
the `DEBUG_JS` variable.