WT-11875: Fix WTreeView size sharing in FlexLayout
When several WTreeView are used inside of Flex layouts, the size of one
of the WTreeView will take as much space as possible, while the other
will take as few space as possible. This is due to WTreeView changing
the size of the element composing it depending on the size it is
allocated, which in turn affect the size it is allocated by the flex
layout.

To fix the issue, WTreeViews will now always measure their size at the
same time.