WT-13666: Deprecate cpp17::any any implementation
Since Wt must now be compiled with C++17 or above, only std::any or
should be used. Therefore, cpp17::any is marked as deprecated and will
be removed in a future version.

We however cannot deprecate the version that of cpp17 that uses std::any
so that there is a way to avoid all of the warnings. We will mark that
as deprecated once the other options are removed.