WT-10923: Fix supportsCookie for WidgetSet
Since setting cookies with SameSite=Lax is impossible cross site, the
test to check cookies availability did not work in WidgetSet mode. This
cannot, however, be changed by simply setting the testcookie attribute
SameSite to None since Chrome-based browser only allows to set such a
cookie over HTTPS with the secure attribute.

For this reason, in WidgetSet mode, a test cookie is set by the request
that serves the Boot.js file.