JWT-14038: Do not propagate visibility or enabled state too early
For a `WCompositeWidget`, the `impl_` carries the actual element that
ought to be visible or enabled. For JWt an additional constructor is
made, that takes a `WContainerWidget` as input, and adds the
`WCompositeWidget` to it. This causes the visibility and enabled state
propagation to run, but without the `WCompositeWidget` actually having
an implementation.

When setting the parent, we now ensure that the `impl_` exists before
propagating.