JWT-14051: Fix makeAbsoluteUrl() dots handling
A trick used in WebSession::makeAbsoluteUrl() to remove the starting
dot of the url was not correctly being translated to java. URLs starting
with a dot would instead be reduced to there first character after that
dot, and an index out of bound exception was raised if the URL was only
composed of a dot.