WT-13861: Delay the visual `WStackedWidget` current widget selection
When setting the `WStackedWidget` current widget, this action is
performed in JavaScript. Where the client side manages the visibility of
the current item.

Due to previous changes with menus (combo of WT-11868, WT-13291, and,
WT-13585), a case was introduced where a lazily loaded item would be
incompletely scaled to fit its layout.

This happened because the `WMenu` (or `WTabWidget`) selection is
attached to prelearned slots. This slot will set the currently selected
widget in the `WStackedWidget`. The call to do this, `setCurrent` will
perform a layout adjust (resize), and this carries a non-stateless
side-effect.

Since the `setCurrent` call is not completely "stateless", it is wrapped
in a timeout. This way the JavaScript will go over the full server
response, and THEN perform the current widget switch in the stack.

The old active stack widget is still immediately set to `hidden`, and
the content is added to it. The final layout change is then kept for the
function called from the timeout.