WT-12407: Fix submenu selecting themself on clear
To fix WT-8135, submenus select themself when select is called on them.
This makes sense when select is called to select one of their menu
items, but not when it is called with an index of -1 as it is supposed
to remove their current selection.

This ensures that clearing the selection of a submenu does not select
it.