WT-12463: Fix WStreamResource concurent get issue
Fix WStreamResource not correctly handling several get requests with
different ranges at the same time. This was due to WStreamResource
only keeping in memory the end index of the last get request.

This could lead to Wt sending to few or too much data to some clients,
as well as other potential race conditions since the end index was not
protected by any mutex.

This is solved by storing the end index inside of the
Http::ResponseContinuation, as proposed by Dries Mys.