WT-14079: Correctly add valid `WStandardItem`s when inserting items
When inserting new columns or rows, we will construct default
`WStandardItem` elements, so that their content can be updated, and they
can be used by views.

Before these were `nullptr`s, and editing newly added items would result
in `nullptr` exceptions.

The items are then properly adopted by their parent, and added to the
model if this is applicable.