WT-13798: Fix popup submenu auto-hide
Fix WPopupMenu always using the auto-hide option of the top-level popup
menu, meaning that setting this option on a submenu of that popup would
not be taken into account.

This allows to autoclose a popup submenu even if the top-level popup
menu does not. The opposite is not true as closing a menu closes its
decendent.