WT-14094: Fix width of WTableView contents
The width of the WTableView content was set as the size of the header
container. However, when there was a vertical scrollbar, the margin
added at the end of the header container was not taken into account.

To solve the issue, the width of the content is now computed similarly
to how the height is computed.