WT-13798: Add auto-hide behaviours for WPopupMenu
The way WPopupMenu handles auto-hide was lacking some customizability.
For instance, it was not possible to only close the last opened submenu
when the mouse leaves the menu. This could be annoying in cases where
a submenu hides its parent menu due to lack of space.

This add WPopupMenu::setAutoHideBehaviour() which allows to change how
WPopupMenu handles auto-hide.