WT-13777: Check the bounds of `WTreeView` model removals
When the model of a `WTreeView` is changed, this change is propagated to
the view, so that it can update its expanded/collapsed items, and
correctly map items, omitting the removed items.

In this functionality, no bounds checking occurred, allowing the model
to remove non-existing items. In some cases this could lead to a
segfault.

When removals occur now, the range to be removed, is verified against
the current state of the model, and eventual offending cases are
corrected to be within the model's range.