WT-13907: Fix TinyMCE wrong version warning
The version of TinyMCE was checked on the server side, expecting the
file to be available on the server. It is however possible to specify an
URL for TinyMCE that is not handled by the server.

This changes it so that the check is done on the client side, once the
exact version of TinyMCE is known.