WT-14078: Fix WDialog movement in Bootstrap 5 RTL
Bootstrap 5 sets the left (in LTR) attribute of modal element to 0px.
In Bootstrap 5 RTL this is changed to right. Since Wt was using the left
attribute to move and center WDialog, the value set for the right
attribute by Bootstrap 5 RTL was never overwritten, making the WDialog
stick to the right side of the window.