WT-14192: Make TOTP `validateCode()` accept one future token
The implementation for `validateCode()` used to validate the token
against two tokens. One generated now, and one generated in the previous
period. This is to ensure that a user's delayed action when copying the
code over, and entering it into the input field is still accepted.

Now, we notice that when the clock of two devices is out of sync, we
could get a token that is essentially a "future" token. This is not
accepted. However, if the user then retries the same token, it is
accepted a couple of seconds later.

OWASP notes that both a token from the "past" and "future" can be
accepted for a smoother experience:

https://owasp.org/www-project-web-security-testing-guide/latest/4-Web_Application_Security_Testing/04-Authentication_Testing/11-Testing_Multi-Factor_Authentication#hotp-and-totp