WT-3824: Make `WMessageResources` thread-safe
Any changes to its `resources_` object are safely put behind a mutex.
This ensures that all changes are serialized, and no concurrent access
or writes happen.

This also ensures that all instances where a `WMessageResources` is
shared (in the case of a shared `WLocalizedString` object, which can be
assigned to a `WApplication` or `WServer`) this can safely occur.